Added subproject test case. It also tests static library creation indirectly.

+++ tests/runTests.sh
@@ -3,18 +3,24 @@
 # BashBuild test runner
 #
 
-CONFIGSCRIPT="$(echo $(which $0) | sed 's@tests/runTests.sh@configure@')"
+CONFIGSCRIPT="$(readlink -f $(dirname $0)/../configure)"
 
 function runTest () {
     cd "$1"
     $CONFIGSCRIPT reconfigure cfgVars > configure
-    for i in $2; do $CONFIGSCRIPT reconfigure cfgVars > configure; done;
-    chmod +x configure
-    ./configure && make > /dev/null
+    for i in $2; do $CONFIGSCRIPT reconfigure $i/cfgVars > $i/configure; done;
+
+    sh configure
+    if [ ! -f 'Makefile' ]; then echo "Failed test '$1'"; exit; fi;
+
+    make > /dev/null
+    if [ ! -f "$1" ]; then echo "Failed test '$1'"; exit; fi;
     if [ ! "$(./$1)" == 'Success!' ]; then echo "Failed test '$1'"; exit; fi;
+
     make distclean > /dev/null
     rm configure
     cd ..
 }
 
 runTest 'basic' ''
+runTest 'subProject' 'libhello'
